chore: Move guest_vm_runner to a separate crate
