+++ ic-os/components/hostos-scripts/guestos/guestos.service
@@ -10,7 +10,7 @@ RequiresMountsFor=/var
 [Service]
 Type=notify
 ExecStartPre=/opt/ic/bin/detect-first-boot.sh
-ExecStart=/opt/ic/bin/hostos_tool run-guest-vm
+ExecStart=/opt/ic/bin/guest_vm_runner
 ExecStartPost=/opt/ic/bin/manageboot.sh hostos confirm
 Restart=always
 RestartSec=60
