+++ ic-os/hostos/defs.bzl
@@ -31,6 +31,7 @@ def image_deps(mode, _malicious = False):
             # additional files to install
             "//rs/ic_os/release:vsock_host": "/opt/ic/bin/vsock_host:0755",
             "//rs/ic_os/release:hostos_tool": "/opt/ic/bin/hostos_tool:0755",
+            "//rs/ic_os/release:guest_vm_runner": "/opt/ic/bin/guest_vm_runner:0755",
             "//rs/ic_os/release:metrics-proxy": "/opt/ic/bin/metrics-proxy:0755",
             "//rs/ic_os/release:config": "/opt/ic/bin/config:0755",
 
@@ -75,8 +76,12 @@ def image_deps(mode, _malicious = False):
     if "dev" in mode:
         deps["rootfs"].pop("//rs/ic_os/release:config", None)
         deps["rootfs"].update({"//rs/ic_os/release:config_dev": "/opt/ic/bin/config:0755"})
+
         deps["rootfs"].pop("//rs/ic_os/release:hostos_tool", None)
         deps["rootfs"].update({"//rs/ic_os/release:hostos_tool_dev": "/opt/ic/bin/hostos_tool:0755"})
+
+        deps["rootfs"].pop("//rs/ic_os/release:guest_vm_runner", None)
+        deps["rootfs"].update({"//rs/ic_os/release:guest_vm_runner_dev": "/opt/ic/bin/guest_vm_runner:0755"})
 
     return deps
