+++ rs/ic_os/config/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2021"
 
 [dependencies]
 anyhow = { workspace = true }
-askama = { workspace = true }
 ic-types = { path = "../../types/types" }
 clap = { workspace = true }
 macaddr = { workspace = true }
@@ -24,7 +23,6 @@ tempfile = { workspace = true }
 once_cell = "1.8"
 tempfile = { workspace = true }
 url = { workspace = true }
-goldenfile = { workspace = true }
 
 [lib]
 name = "config"
