+++ rs/ic_os/config/src/lib.rs
@@ -1,7 +1,6 @@
 pub mod config_ini;
 pub mod deployment_json;
 pub mod generate_testnet_config;
-pub mod guest_vm_config;
 pub mod guestos_bootstrap_image;
 pub mod guestos_config;
