+++ rs/ic_os/os_tools/guest_vm_runner/golden/guestos_vm_kvm.xml
