+++ rs/ic_os/os_tools/guest_vm_runner/templates/guestos_vm_template.xml
