+++ rs/ic_os/release/BUILD.bazel
@@ -5,7 +5,9 @@ package(default_visibility = ["//rs:ic-os-pkg"])
 OBJECTS = {
     "fstrim_tool": "//rs/ic_os/fstrim_tool:fstrim_tool_bin",
     "guestos_tool": "//rs/ic_os/os_tools/guestos_tool:guestos_tool",
-    "hostos_tool": "//rs/ic_os/os_tools/hostos_tool:hostos_tool",
+    "guest_vm_runner": "//rs/ic_os/os_tools/guest_vm_runner",
+    "guest_vm_runner_dev": "//rs/ic_os/os_tools/guest_vm_runner:guest_vm_runner_dev",
+    "hostos_tool": "//rs/ic_os/os_tools/hostos_tool",
     "hostos_tool_dev": "//rs/ic_os/os_tools/hostos_tool:hostos_tool_dev",
     "nft-exporter": "//rs/ic_os/nft_exporter:nft-exporter",
     "setupos_tool": "//rs/ic_os/os_tools/setupos_tool:setupos_tool",
